requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
python-json-logger==2.0.7
requests==2.31.0
python-dotenv==1.0.0
google-genai==1.0.0
python-pptx==0.6.23
playwright==1.40.0
Pillow==10.1.0
